+++ src/index.js
@@ -2,7 +2,7 @@ import dotenv from "dotenv";    //.env 관리 모듈
 import cors from "cors"   //cors 관리 모듈
 import express from "express";
 import { handleUserMisAdd, handleUserSignUp } from "./controllers/user.controller.js";
-import { handleReviewWrite, handleStoreAdd, handleStoreMisAdd } from "./controllers/store.controller.js";
+import { handleListStoreReviews, handleReviewWrite, handleStoreAdd, handleStoreMisAdd } from "./controllers/store.controller.js";
 
 dotenv.config();    //config()를 호출해 env에 있는 내용 접근
 
@@ -27,6 +27,8 @@ app.post("/api/v1/users/missions", handleUserMisAdd);
 app.post("/api/v1/stores", handleStoreAdd);
 app.post("/api/v1/stores/reviews", handleReviewWrite);
 app.post("/api/v1/stores/missions", handleStoreMisAdd);
+app.get("/api/v1/stores/:storeId/reviews", handleListStoreReviews);
+
 app.listen(port, () => {
   console.log(`Example app listening on port ${port}`);
 });
